feat: Handle `ChatContainerLoader` into `ChatContainer`
