+++ src/hooks/matrix/useGlobalEventListeners.ts
@@ -29,7 +29,7 @@ const useGlobalEventListeners = () => {
       return false
     }
 
-    return getNotificationsData().some(notification => notification.isRead)
+    return getNotificationsData().some(notification => !notification.isRead)
   }, [state])
 
   useEventListener(
